Diagrams for OBIB and others
These are the diagrams in development for representing paradigmatic cases, testing the boundaries of our ontological representation capabilities. The 'Diagrams of Processes' has much more focus on talking about what is involved in the planned process 'informed consent' in different contexts. The 'Diagrams of Use Cases' has specific scenarios that we are striving to be able to handle with our project.